+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Arbitrary.hs
@@ -55,6 +55,7 @@ import Cardano.Ledger.BaseTypes (
   Exclusive (..),
   HasZero,
   Inclusive (..),
+  Milliseconds32 (..),
   Mismatch (..),
   Network (..),
   NonNegativeInterval,
@@ -186,6 +187,9 @@ instance Arbitrary Url where
 
 instance Arbitrary Port where
   arbitrary = fromIntegral @Word16 @Port <$> arbitrary
+
+instance Arbitrary Milliseconds32 where
+  arbitrary = Milliseconds32 <$> arbitrary
 
 -- JSON instances can't roundtrip, unless these are decimal.