+++ eras/dijkstra/impl/CHANGELOG.md
@@ -57,6 +57,13 @@
   - Add `hkdMinPoolMarginL` to `DijkstraEraPParams`
   - Add `ppMinPoolMarginL` and `ppuMinPoolMarginL`
   - Override `ppMinPoolMarginG` in `EraPParams` instance
+* Add the nine Leios protocol parameters of CIP-164:
+  - Add `dppLeiosAnnouncementPeriodLength`, `dppLeiosVotePeriodLength`,
+    `dppLeiosDiffusionPeriodLength`, `dppLeiosCommitteeSize`, `dppLeiosQuorumStakeThreshold`,
+    `dppMaxEndorserBlockReferencesSize`, `dppMaxEndorserBlockTxsSize`,
+    `dppMaxEndorserBlockExUnits`, `dppMaxRefScriptSizePerEndorserBlock` to `DijkstraPParams`
+  - Add corresponding fields to `UpgradeDijkstraPParams`
+  - Add corresponding HKD lenses, pp/ppu lenses to `DijkstraEraPParams`
 
 ### `cddl`
 
@@ -64,6 +71,7 @@
 * Replace the `transaction` and `transaction_mempool` rules with `block_transaction` and `mempool_transaction`
 * Remove the `invalid_transactions` rule and drop the field from `block_body`
 * Add `HuddleRule "vrf_cert"` instance
+* Add Leios protocol parameter entries (tags 40-48) in `protocol_param_update`
 * Add `max_pledge_leverage` rule and its entry in `protocol_param_update`
 
 ### `testlib`